Build project components and  styles
Organized project components into separate Vue files for better code organization and readability.
Reorganized CSS styles to improve consistency and maintainability across the project.

-Updated image paths to ensure proper rendering and accessibility.
Enhanced form validation logic for better user experience.

-Implemented animations for error handling to provide visual feedback to users.

-Standardized class naming conventions for easier collaboration and future updates.

-Ensured responsiveness and accessibility considerations were maintained throughout the codebase.

-Applied best practices for Vue.js development to optimize performance and maintainability.
