+++ package.json
@@ -8,6 +8,8 @@
   },
   "dependencies": {
     "core-js": "^3.8.3",
+    "pdfvuer": "^2.0.1",
+    "pspdfkit": "^2024.2.0",
     "vue": "^3.2.13",
     "vue-router": "^4.0.3"
   },
