+++ src/assets/project_img/cahier_des_charges/cdc_sample.png
