+++ src/assets/project_img/comments/comment_sample.png
