+++ src/data/projects_data.js
@@ -1,0 +1,34 @@
+
+ 
+
+export const projectsData = 
+
+{
+    myCv : {
+        id : 1 ,
+        title : "Mon CV En HTML Et CSS",
+        date : "11/02/2024",
+        technologies : ["HTML", "CSS"],
+        imgSrc : [require("@/assets/project_img/cv/cv_1.png"),require("../assets/project_img/cv/cv_2.png"), require("../assets/project_img/cv/cv_3.png")],
+        link : "https://github.com/IekoTsu/Mon-CV"
+    },
+
+    cahierDeCharge : {
+        id : 2 ,
+        title : "Cahier des charges",
+        date : "16/02/2024",
+        technologies : ["Adobe acrobat", "Adobe Photoshop", "PDF"],
+        imgSrc : [require("../assets/project_img/cahier_des_charges/cdc_1.png"), require("../assets/project_img/cahier_des_charges/cdc_2.png"), require("../assets/project_img/cahier_des_charges/cdc_3.png") ],
+        link : "https://drive.google.com/file/d/1NNaqic0aL_GNN3yCPSKj6Im59B1sWcp_/view?usp=sharing"
+    },
+
+    commentField : {
+        id : 3 ,
+        title : "Espace de commentaire",
+        date : "11/03/2024",
+        technologies : ["HTML", "CSS", "JavaScript" ],
+        imgSrc : [require("../assets/project_img/comments/comment_1.png"), require("../assets/project_img/comments/comment_2.png"), require("../assets/project_img/comments/comment_3.png"), require("../assets/project_img/comments/comment_4.png")],
+        link : "https://github.com/IekoTsu/Un_Espace_De_Commentaires_Dynamique"
+
+    }
+}
