+++ src/views/notFound.vue
@@ -1,0 +1,12 @@
+<template>
+    <img class="notFound" src="../assets/decorations/404.png">
+    <div class="error">
+        <h2>404</h2>
+        <h3>page Not Found</h3>
+    </div>
+</template>
+<style scoped>
+    img{
+        padding-top: 10rem;
+    }
+</style>
